requirements.txt
agno==1.5.10
chromadb==0.4.24
pypdf==3.17.*
python-docx==1.1.*
sentence-transformers==2.7.*
streamlit==1.35.*
python-dotenv==1.0.*
google-generativeai==0.5.*
huggingface-hub
torch
numpy<2.0.0
pysqlite3-binary
